Optimize Ccheckbound during cmmgen.
Attempt to replace Cmm constructs like

  (checkbound (>>u arg n) m)

where n and m are integer constants with

  (checkbound arg (m << m + (1 << n - 1)))

which is both shorter and results in faster code. On amd64 and i386 this
replaces a sequence of mov,shr,cmp,jbe with a sequence of mov,cmp,jbe.
This could be further optimized to a sequence of cmp,jbe in the backends.