Optimize common case of Ccheckbound for amd64.
One common case of Ccheckbound on amd64 (considering the optimization
from the previous commit), is:

  Ccheckbound, [Cop(Cload _, [loc]); Cconst_int n]

Using the normal code selection, this generates a sequence of mov,cmp,jbe.
But the cmp instruction supports memory operands with immediate operands,
which would be a sequence of cmp,jbe. So we add an Ispecific operation
Icheckboundmem_imm, which does exactly this and is selected appropriately.