Improve Cmm code generator for "for" loops.
The code generated by cmmgen for Ufor loops doesn't optimize in case of
constant low and high expressions. This is fixed by a special case where
the initial comparison can be skipped in case of constant low and high
subexpressions.

In addition, this commit also reorders the generated code by moving the
initial comparison out of the catch block in the general case. This may
enable certain optimizations in later phases.