Fix interval construction.
This patch introduces a new mechanism to handle overlapping live intervals
for arguments/results, which is no longer based on the ideas used in the
Java HotSpot Client Compiler as described by Wimmer. The new approach is
superior, easier to understand and fixes a nasty bug in the case of ARM
code generation.

Signed-off-by: Benedikt Meurer <benedikt.meurer@googlemail.com>